requirements.txt
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
ebaysdk==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
